RAC-313: add test
