+++ src/Akeneo/Platform/Bundle/UIBundle/Resources/workspaces/shared/tests/front/unit/hooks/useDocumentVisibility.unit.ts
@@ -1,0 +1,19 @@
+import {useDocumentVisibility} from '../../../../src/hooks/useDocumentVisibility';
+import {renderHookWithProviders} from '../utils';
+
+test('It return the visibility of the document', () => {
+  const {result} = renderHookWithProviders(() => useDocumentVisibility());
+
+  expect(result.current).toEqual(true);
+});
+
+test('It return false when the document is not visible', async () => {
+  const {result, waitForNextUpdate} = renderHookWithProviders(() => useDocumentVisibility());
+
+  Object.defineProperty(document, 'visibilityState', {value: 'hidden'});
+  document.dispatchEvent(new Event('visibilitychange'));
+
+  await waitForNextUpdate();
+
+  expect(result.current).toEqual(false);
+});
